+++ repexecutor/src/main/java/com/provys/report/jooxml/datasource/RootDataRecord.java
@@ -21,8 +21,8 @@ public class RootDataRecord extends DataRecordAncestor {
     }
 
     @Override
-    public int getRowNumber() {
-        return 0;
+    public long getRowNumber() {
+        return 0L;
     }
 
     @Nonnull
